+++ neoforge/src/main/java/dev/terminalmc/clientsort/client/ClientSortNeoForge.java
@@ -22,16 +22,12 @@ import dev.terminalmc.clientsort.client.util.KeybindManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.server.packs.PackLocationInfo;
-import net.minecraft.server.packs.PackSelectionConfig;
+import net.minecraft.resources.Identifier;
 import net.minecraft.server.packs.PackType;
-import net.minecraft.server.packs.PathPackResources;
-import net.minecraft.server.packs.repository.Pack;
 import net.minecraft.server.packs.repository.Pack.Position;
 import net.minecraft.server.packs.repository.PackSource;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.ModList;
 import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
@@ -41,9 +37,6 @@ import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 import net.neoforged.neoforge.event.AddPackFindersEvent;
 import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
 import net.neoforged.neoforge.network.registration.PayloadRegistrar;
-import net.neoforged.neoforgespi.locating.IModFile;
-
-import java.util.Optional;
 
 import static dev.terminalmc.clientsort.util.Localization.localized;
 
@@ -82,32 +75,17 @@ public class ClientSortNeoForge {
     @SubscribeEvent
     static void addPackFinders(AddPackFindersEvent event) {
         if (event.getPackType().equals(PackType.CLIENT_RESOURCES)) {
-            event.addRepositorySource(((packConsumer) -> {
-                String packId = "clientsort-dark-mode";
-                IModFile file = ModList.get().getModFileById(ClientSort.MOD_ID).getFile();
-                try {
-                    Pack.ResourcesSupplier supplier = new PathPackResources.PathResourcesSupplier(
-                            file.getFilePath().resolve("resourcepacks/" + packId)
-                    );
-                    PackLocationInfo info = new PackLocationInfo(
-                            ClientSort.MOD_ID + ":" + packId,
-                            localized("resourcepack", "dark-mode"),
-                            PackSource.BUILT_IN,
-                            Optional.empty()
-                    );
-                    Pack pack = Pack.readMetaAndCreate(
-                            info,
-                            supplier,
-                            PackType.CLIENT_RESOURCES,
-                            new PackSelectionConfig(false, Position.TOP, false)
-                    );
-                    if (pack != null) {
-                        packConsumer.accept(pack);
-                    }
-                } catch (NullPointerException e) {
-                    e.fillInStackTrace();
-                }
-            }));
+            event.addPackFinders(
+                    Identifier.fromNamespaceAndPath(
+                            ClientSort.MOD_ID,
+                            "resourcepacks/clientsort-dark-mode"
+                    ),
+                    PackType.CLIENT_RESOURCES,
+                    localized("resourcepack", "dark-mode"),
+                    PackSource.BUILT_IN,
+                    false,
+                    Position.TOP
+            );
         }
     }